requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
dj-database-url==2.2.0
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==23.0.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==3.12.3
python-decouple==3.8
pytz==2024.1
requests==2.31.0
setuptools==73.0.1
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
